Break up components into files
